:green_heart: Remove library name from ci.yml
Committing & pushing changes to workflow files requires additional
permissions in github actions. The action can either get elevated
permissions for this task, or the library name can be removed from the
workflow dispatch. Removing the library name means that the package name
and the repo name must match. I personally believe that is an acceptable
constraint. If this needs to be changed, then the library name can be
coded directly.